+++ features/support/env.rb
@@ -5,13 +5,15 @@ require 'require_all'
 require 'simplecov'
 require 'testcentricity_web'
 
-SimpleCov.command_name "features #{ENV['WEB_BROWSER']}" + (ENV['TEST_ENV_NUMBER'] || '')
+include TestCentricity
+
+SimpleCov.command_name("features-#{ENV['WEB_BROWSER']}-#{ENV['SELENIUM']}" + (ENV['TEST_ENV_NUMBER'] || ''))
 
 # require_relative 'world_data'
 require_relative 'world_pages'
 
 # require_rel 'data'
-# require_rel 'sections'
+require_rel 'sections'
 require_rel 'pages'
 
 $LOAD_PATH << './lib'
@@ -27,8 +29,9 @@ I18n.locale = ENV['LOCALE']
 Faker::Config.locale = ENV['LOCALE']
 
 # instantiate all data objects and target test environment
-# include WorldData
-# environs.find_environ(ENV['TEST_ENVIRONMENT'], :yaml)
+include WorldData
+ENV['DATA_SOURCE'] = 'yaml' unless ENV['DATA_SOURCE']
+environs.find_environ(ENV['TEST_ENVIRONMENT'], ENV['DATA_SOURCE'].downcase.to_sym)
 # WorldData.instantiate_data_objects
 
 # instantiate all page objects
@@ -38,6 +41,10 @@ WorldPages.instantiate_page_objects
 # establish connection to WebDriver and target web browser
 Webdrivers.cache_time = 86_400
 
-options = { app_host: "file://#{File.dirname(__FILE__)}/../../test_site" }
-options = { app_host: "http://192.168.1.129"}
-TestCentricity::WebDriverConnect.initialize_web_driver(options)
+
+url = if ENV['TEST_ENVIRONMENT'].downcase == 'local'
+        "file://#{File.dirname(__FILE__)}/../../test_site"
+      else
+        Environ.current.app_host
+      end
+WebDriverConnect.initialize_web_driver(app_host: url)